rename iavl-v1.2 patches -> iavl-v1.2-kava-v0.26
this is in order to distinguish the patches from potential future forked
iavl v1 changesets.

particularly, we expect to drop 0001 at the next major upgrade for easier
future compatibility with upstream.
(it is only required for backwards-compatibility with v0)

of course, ideally, we get to just use an unpatched upstream 🙂